Add DuckDuckGo support
Note: DuckDuckGo restricts its API-based searches to its 'Instant Answer API' which is essentially a slimmed down search engine that returns extremely basic results. Short of pretending to be a browser and search their main search and then parsing the data (which would violate their terms and conditions) this is the only way to retrieve results from them.

- Fixed bad characters showing in descriptions

Signed-off-by: Steven Downer <grinch@outlook.com>